app/requirements.txt
uvicorn>=0.23.0
openai>=1.0.0
fastapi>=0.100.0
pydantic>=1.10.0,<3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
websockets>=12.0
slowapi>=0.1.9